requirements.txt
slack_bolt~=1.18.1
python-dotenv~=1.0.1
requests~=2.31.0
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.30 
flask~=3.0.3
gunicorn~=22.0.0